pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rre-burp"
version = "1.0.0"
description = "Request Response Explorer with AI Integration - Intelligent web traffic analysis toolkit"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Farzan Karimi", email = "farzan@farzan.ca"},
]
keywords = ["web-traffic", "har-analysis", "security", "ai", "langchain", "rre"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Security",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "langchain>=0.1.0",
    "langchain-openai>=0.1.0",
    "openai>=1.0.0",
    "playwright>=1.45.0",
    "PyYAML>=6.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "colorama>=0.4.6",
    "rich>=13.0.0",
    "fastapi>=0.116.1",
    "uvicorn[standard]>=0.35.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
]

[project.scripts]
site-explorer = "src.site_explorer_cli:main"

[project.urls]
Homepage = "https://github.com/farzan/rre-burp"
Repository = "https://github.com/farzan/rre-burp"
Documentation = "https://github.com/farzan/rre-burp#readme"
Issues = "https://github.com/farzan/rre-burp/issues"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    ".venv",
    "build",
    "dist",
    "*.egg-info"
]

